drop unused code
