+++ src/kernels/execution/cellExecutionMessageHandler.ts
@@ -34,7 +34,8 @@ import {
     traceCellMessage,
     cellOutputToVSCCellOutput,
     translateCellDisplayOutput,
-    CellOutputMimeTypes
+    CellOutputMimeTypes,
+    parseStackTrace as findErrorLocation
 } from './helpers';
 import { swallowExceptions } from '../../platform/common/utils/decorators';
 import { noop } from '../../platform/common/utils/misc';
@@ -43,7 +44,6 @@ import { handleTensorBoardDisplayDataOutput } from './executionHelpers';
 import { Identifiers, WIDGET_MIMETYPE } from '../../platform/common/constants';
 import { CellOutputDisplayIdTracker } from './cellDisplayIdTracker';
 import { createDeferred } from '../../platform/common/utils/async';
-import { CellFailureDiagnostics } from './cellFailureDiagnostics';
 
 // Helper interface for the set_next_input execute reply payload
 interface ISetNextInputPayload {
@@ -103,7 +103,6 @@ export class CellExecutionMessageHandler implements IDisposable {
      */
     private clearOutputOnNextUpdateToOutput?: boolean;
 
-    private failureDiagostics = new CellFailureDiagnostics();
     private execution?: NotebookCellExecution;
     private readonly _onErrorHandlingIOPubMessage = new EventEmitter<{
         error: Error;
@@ -1079,7 +1078,7 @@ export class CellExecutionMessageHandler implements IDisposable {
             if (cellExecution && msg.content.ename !== 'KeyboardInterrupt') {
                 cellExecution.errorInfo = {
                     message: `${msg.content.ename}: ${msg.content.evalue}`,
-                    location: this.failureDiagostics.parseStackTrace(msg.content.traceback, this.cell).range,
+                    location: findErrorLocation(msg.content.traceback, this.cell),
                     uri: this.cell.document.uri,
                     stack: msg.content.traceback.join('\n')
                 };
